+++ Profiles/Akash.md
@@ -1,0 +1,25 @@
+# Akash Srivastava
+
+### Location
+
+Ayodhya , India
+
+### Academics
+
+JB Acadmey
+
+### Interests
+
+Chess , Acting , Novels 
+
+### Development
+
+MAde some awesome android and python projects do checkout my github 
+
+### Projects
+
+- [Movies App](https://github.com/Akashsri3bi/Movies_App_) Made an Movies application using android , java , and Rest Api
+
+### Profile Link
+
+[Akash Srivastava](https://github.com/Akashsri3bi)
